+++ Sources/Styleguide/EnvironmentValues.Styleguide.swift
@@ -1,0 +1,6 @@
+import SwiftUI
+
+public extension EnvironmentValues {
+	/// The styleguide available to SwiftUI views in the environment.
+	@Entry var styleguide: Styleguide = .default
+}